requirements.txt
discord
python-dotenv
pytube